+++ rust/src/quote/mod.rs
@@ -17,6 +17,9 @@ pub use push_types::{
 };
 pub use sub_flags::SubFlags;
 pub use types::{
+    AssetAllocationGroup,
+    AssetAllocationItem,
+    AssetAllocationResponse,
     Brokers,
     CalcIndex,
     Candlestick,
@@ -25,11 +28,13 @@ pub use types::{
     CapitalFlowLine,
     Depth,
     DerivativeType,
+    ElementType,
     FilingItem,
     FilterWarrantExpiryDate,
     FilterWarrantInOutBoundsType,
     Granularity,
     HistoryMarketTemperatureResponse,
+    HoldingDetail,
     IntradayLine,
     IssuerInfo,
     MarketTemperature,